Relocate recipe related files
