+++ pyodide_build/build_env.py
@@ -16,7 +16,6 @@ from packaging.tags import Tag, compatible_tags, cpython_tags
 from pyodide_build import __version__
 from pyodide_build.common import search_pyproject_toml, to_bool, xbuildenv_dirname
 from pyodide_build.config import ConfigManager, CrossBuildEnvConfigManager
-from pyodide_build.recipe import load_all_recipes
 
 RUST_BUILD_PRELUDE = """
 rustup default ${RUST_TOOLCHAIN}
@@ -188,6 +187,8 @@ def get_hostsitepackages() -> str:
 
 @functools.cache
 def get_unisolated_packages() -> list[str]:
+    # TODO: Remove this function (and use remote package index)
+    # https://github.com/pyodide/pyodide-build/issues/43
     PYODIDE_ROOT = get_pyodide_root()
 
     unisolated_file = PYODIDE_ROOT / "unisolated.txt"
@@ -195,6 +196,8 @@ def get_unisolated_packages() -> list[str]:
         # in xbuild env, read from file
         unisolated_packages = unisolated_file.read_text().splitlines()
     else:
+        from pyodide_build.recipe.loader import load_all_recipes
+
         unisolated_packages = []
         recipe_dir = PYODIDE_ROOT / "packages"
         recipes = load_all_recipes(recipe_dir)
