+++ pyodide_build/cli/build_recipes.py
@@ -4,11 +4,12 @@ from pathlib import Path
 
 import typer
 
-from pyodide_build import build_env, buildall, recipe
+from pyodide_build import build_env
 from pyodide_build.build_env import BuildArgs, init_environment
-from pyodide_build.buildpkg import RecipeBuilder
+from pyodide_build.recipe.buildpkg import RecipeBuilder
 from pyodide_build.common import get_num_cores
 from pyodide_build.logger import logger
+from pyodide_build.recipe.loader import buildall, recipe
 
 
 @dataclasses.dataclass(eq=False, order=False, kw_only=True)
