+++ pyodide_build/recipe/buildpkg.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 """
 Builds a Pyodide package.
 """
