+++ pyodide_build/recipe/mkpkg.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 import contextlib
 import json
 import shutil
